sync

+++ Modules/ModUpdater.cs
@@ -6,40 +6,55 @@ using System.Reflection;
 using System.Threading.Tasks;
 using HarmonyLib;
 using Newtonsoft.Json.Linq;
+using TMPro;
 using Twitch;
 using UnityEngine;
+using UnityEngine.Events;
 using static EHR.Translator;
-
 
 namespace EHR;
 
 [HarmonyPatch]
-public class ModUpdater
+public static class ModUpdater
 {
     private const string URLGithub = "https://api.github.com/repos/Gurge44/EndlessHostRoles";
+    public const bool ForceUpdate = false;
     public static bool HasUpdate;
+    private static bool FirstNotify = true;
     private static bool HasOutdate;
-    public static bool ForceUpdate = false;
     public static bool IsBroken;
     private static bool IsChecked;
     private static Version LatestVersion;
+    private static string LatestTitleModName;
     private static string LatestTitle;
     public static string DownloadUrl;
     private static GenericPopup InfoPopup;
+    private static GenericPopup InfoPopupV2;
+    private static readonly HttpClient HttpClient = new();
 
-    [HarmonyPatch(typeof(MainMenuManager), nameof(MainMenuManager.Start)), HarmonyPrefix]
+    [HarmonyPatch(typeof(MainMenuManager), nameof(MainMenuManager.Start))]
+    [HarmonyPrefix]
     [HarmonyPriority(2)]
     public static void Start_Prefix()
     {
-        NewVersionCheck();
-        DeleteOldFiles();
+        if (!OperatingSystem.IsAndroid())
+        {
+            // Version checks are not handled on Android
+            NewVersionCheck();
+            DeleteOldFiles();
+        }
+        
         InfoPopup = Object.Instantiate(TwitchManager.Instance.TwitchPopup);
         InfoPopup.name = "InfoPopup";
         InfoPopup.TextAreaTMP.GetComponent<RectTransform>().sizeDelta = new(2.5f, 2f);
-        if (!IsChecked)
+
+        InfoPopupV2 = Object.Instantiate(TwitchManager.Instance.TwitchPopup);
+        InfoPopupV2.name = "InfoPopupV2";
+
+        if (!OperatingSystem.IsAndroid() && !IsChecked)
         {
             bool done = CheckReleaseFromGithub(Main.BetaBuildUrl.Value != "").GetAwaiter().GetResult();
-            Logger.Warn("done: " + done, "CheckRelease");
+            Logger.Msg("done: " + done, "CheckRelease");
             Logger.Info("hasupdate: " + HasUpdate, "CheckRelease");
             Logger.Info("forceupdate: " + ForceUpdate, "CheckRelease");
             Logger.Info("downloadUrl: " + DownloadUrl, "CheckRelease");
@@ -47,36 +62,47 @@ public class ModUpdater
         }
     }
 
+    public static void ShowAvailableUpdate()
+    {
+        if (FirstNotify && HasUpdate)
+        {
+            FirstNotify = false;
+            
+            if (!string.IsNullOrWhiteSpace(LatestTitleModName))
+                ShowPopupWithTwoButtons(string.Format(GetString("NewUpdateAvailable"), LatestTitleModName), GetString("updateButton"), onClickOnFirstButton: () => StartUpdate(DownloadUrl, true));
+        }
+    }
+
     public static string Get(string url)
     {
         string result;
-        HttpClient req = new();
-        var res = req.GetAsync(url).Result;
+        HttpResponseMessage res = HttpClient.GetAsync(url).Result;
         Stream stream = res.Content.ReadAsStreamAsync().Result;
+
         try
         {
             using StreamReader reader = new(stream);
             result = reader.ReadToEnd();
         }
-        finally
-        {
-            stream.Close();
-        }
+        finally { stream.Close(); }
 
         return result;
     }
 
     public static async Task<bool> CheckReleaseFromGithub(bool beta = false)
     {
-        Logger.Warn("Checking GitHub Release", "CheckRelease");
+        Logger.Msg("Checking GitHub Release", "CheckRelease");
         const string url = URLGithub + "/releases/latest";
+
         try
         {
             string result;
+
             using (HttpClient client = new())
             {
                 client.DefaultRequestHeaders.Add("User-Agent", "EHR Updater");
-                using var response = await client.GetAsync(new Uri(url), HttpCompletionOption.ResponseContentRead);
+                using HttpResponseMessage response = await client.GetAsync(new Uri(url), HttpCompletionOption.ResponseContentRead);
+
                 if (!response.IsSuccessStatusCode)
                 {
                     Logger.Error($"Response Status Code: {response.StatusCode}", "CheckRelease");
@@ -87,6 +113,9 @@ public class ModUpdater
             }
 
             JObject data = JObject.Parse(result);
+
+            LatestTitleModName = data["name"].ToString();
+
             if (beta)
             {
                 LatestTitle = data["name"].ToString();
@@ -97,10 +126,11 @@ public class ModUpdater
             {
                 LatestVersion = new(data["tag_name"]?.ToString().TrimStart('v') ?? string.Empty);
                 LatestTitle = $"Ver. {LatestVersion}";
-                JArray assets = data["assets"].Cast<JArray>();
-                for (int i = 0; i < assets.Count; i++)
+                var assets = data["assets"].CastFast<JArray>();
+
+                for (var i = 0; i < assets.Count; i++)
                 {
-                    if (assets[i]["name"].ToString() == $"EHR.v{LatestVersion}.zip")
+                    if (assets[i]["name"].ToString() == $"EHR.v{LatestVersion}_Steam.zip")
                     {
                         DownloadUrl = assets[i]["browser_download_url"].ToString();
                         break;
@@ -118,7 +148,7 @@ public class ModUpdater
             Logger.Info("latestVersionl: " + LatestVersion, "GitHub");
             Logger.Info("latestTitle: " + LatestTitle, "GitHub");
 
-            if (string.IsNullOrEmpty(DownloadUrl))
+            if (string.IsNullOrWhiteSpace(DownloadUrl))
             {
                 Logger.Error("No Download URL", "CheckRelease");
                 return false;
@@ -139,6 +169,7 @@ public class ModUpdater
 
     public static void StartUpdate(string url, bool github)
     {
+        if (OperatingSystem.IsAndroid()) return;
         ShowPopup(GetString("updatePleaseWait"), StringNames.Cancel, true, false);
         _ = !github ? DownloadDLL(url) : DownloadDLLGithub(url);
     }
@@ -147,11 +178,11 @@ public class ModUpdater
     {
         try
         {
-            if (Directory.Exists("TOH_DATA") && File.Exists("./EHR_DATA/BanWords.txt"))
+            if (Directory.Exists($"{Main.DataPath}/TOH_DATA") && File.Exists($"{Main.DataPath}/EHR_DATA/BanWords.txt"))
             {
-                DirectoryInfo di = new("TOH_DATA");
+                DirectoryInfo di = new($"{Main.DataPath}/TOH_DATA");
                 di.Delete(true);
-                Logger.Warn("Directory deleted：TOH_DATA", "NewVersionCheck");
+                Logger.Warn("Directory deleted: TOH_DATA", "NewVersionCheck");
             }
         }
         catch (Exception ex)
@@ -167,9 +198,11 @@ public class ModUpdater
     {
         string path = Path.GetDirectoryName(Assembly.GetExecutingAssembly().Location);
         const string searchPattern = "EHR.dll*";
+
         if (path != null)
         {
             string[] files = Directory.GetFiles(path, searchPattern);
+
             try
             {
                 foreach (string filePath in files)
@@ -181,10 +214,7 @@ public class ModUpdater
                     }
                 }
             }
-            catch (Exception e)
-            {
-                Logger.Error($"Failed to clear update residue\n{e}", "DeleteOldFiles");
-            }
+            catch (Exception e) { Logger.Error($"Failed to clear update residue\n{e}", "DeleteOldFiles"); }
         }
     }
 
@@ -195,36 +225,22 @@ public class ModUpdater
             const string savePath = "BepInEx/plugins/EHR.dll.temp";
 
             // Delete the temporary file if it exists
-            if (File.Exists(savePath))
-            {
-                File.Delete(savePath);
-            }
+            if (File.Exists(savePath)) File.Delete(savePath);
 
-            HttpResponseMessage response;
+            HttpResponseMessage response = await HttpClient.GetAsync(url);
 
-            using (HttpClient client = new())
-            {
-                response = await client.GetAsync(url);
-            }
+            if (response is not { IsSuccessStatusCode: true }) throw new($"File retrieval failed with status code: {response.StatusCode}");
 
-            if (response is not { IsSuccessStatusCode: true })
+            await using (Stream stream = await response.Content.ReadAsStreamAsync())
+            await using (var fileStream = new FileStream(savePath, FileMode.Create, FileAccess.Write, FileShare.None, 4096, true))
             {
-                throw new($"File retrieval failed with status code: {response.StatusCode}");
-            }
-
-            await using (var stream = await response.Content.ReadAsStreamAsync())
-            await using (var fileStream = new FileStream(savePath, FileMode.Create, FileAccess.Write, FileShare.None, bufferSize: 4096, useAsync: true))
-            {
-                byte[] buffer = new byte[1024];
+                var buffer = new byte[1024];
                 int length;
 
-                while ((length = await stream.ReadAsync(buffer)) != 0)
-                {
-                    await fileStream.WriteAsync(buffer.AsMemory(0, length));
-                }
+                while ((length = await stream.ReadAsync(buffer)) != 0) await fileStream.WriteAsync(buffer.AsMemory(0, length));
             }
 
-            var fileName = Assembly.GetExecutingAssembly().Location;
+            string fileName = Assembly.GetExecutingAssembly().Location;
             File.Move(fileName, fileName + ".bak");
             File.Move(savePath, fileName);
             ShowPopup(GetString("updateRestart"), StringNames.Close, true);
@@ -246,37 +262,26 @@ public class ModUpdater
             const string savePath = "BepInEx/plugins/EHR.dll.temp";
 
             // Delete the temporary file if it exists
-            if (File.Exists(savePath))
-            {
-                File.Delete(savePath);
-            }
+            if (File.Exists(savePath)) File.Delete(savePath);
 
-            HttpResponseMessage response;
+            HttpResponseMessage response = await HttpClient.GetAsync(url);
 
-            using (HttpClient client = new())
-            {
-                response = await client.GetAsync(url);
-            }
+            if (response is not { IsSuccessStatusCode: true }) throw new($"File retrieval failed with status code: {response.StatusCode}");
 
-            if (response is not { IsSuccessStatusCode: true })
-            {
-                throw new($"File retrieval failed with status code: {response.StatusCode}");
-            }
-
-            await using (var stream = await response.Content.ReadAsStreamAsync())
+            await using (Stream stream = await response.Content.ReadAsStreamAsync())
             using (var archive = new ZipArchive(stream, ZipArchiveMode.Read))
             {
                 // Specify the relative path within the ZIP archive where "EHR.dll" is located
                 const string entryPath = "BepInEx/plugins/EHR.dll";
-                var entry = archive.GetEntry(entryPath) ?? throw new($"'{entryPath}' not found in the ZIP archive");
+                ZipArchiveEntry entry = archive.GetEntry(entryPath) ?? throw new($"'{entryPath}' not found in the ZIP archive");
 
                 // Extract "EHR.dll" to the temporary file
-                await using var entryStream = entry.Open();
-                await using var fileStream = new FileStream(savePath, FileMode.Create, FileAccess.Write, FileShare.None, bufferSize: 4096, useAsync: true);
+                await using Stream entryStream = entry.Open();
+                await using var fileStream = new FileStream(savePath, FileMode.Create, FileAccess.Write, FileShare.None, 4096, true);
                 await entryStream.CopyToAsync(fileStream);
             }
 
-            var fileName = Assembly.GetExecutingAssembly().Location;
+            string fileName = Assembly.GetExecutingAssembly().Location;
             File.Move(fileName, fileName + ".bak");
             File.Move(savePath, fileName);
             ShowPopup(GetString("updateRestart"), StringNames.Close, true);
@@ -293,18 +298,88 @@ public class ModUpdater
 
     public static void ShowPopup(string message, StringNames buttonText, bool showButton = false, bool buttonIsExit = true)
     {
-        if (InfoPopup == null) return;
+        if (!InfoPopup) return;
 
         InfoPopup.Show(message);
-        var button = InfoPopup.transform.FindChild("ExitGame");
-        if (button != null)
+        Transform button = InfoPopup.transform.FindChild("ExitGame");
+
+        if (button)
         {
             button.gameObject.SetActive(showButton);
-            button.GetChild(0).GetComponent<TextTranslatorTMP>().TargetText = buttonText;
-            button.GetChild(0).GetComponent<TextTranslatorTMP>().ResetText();
-            button.GetComponent<PassiveButton>().OnClick = new();
-            if (buttonIsExit) button.GetComponent<PassiveButton>().OnClick.AddListener((Action)Application.Quit);
-            else button.GetComponent<PassiveButton>().OnClick.AddListener((Action)(() => InfoPopup.Close()));
+            var textTranslatorTMP = button.GetChild(0).GetComponent<TextTranslatorTMP>();
+            textTranslatorTMP.TargetText = buttonText;
+            textTranslatorTMP.ResetText();
+            var passiveButton = button.GetComponent<PassiveButton>();
+            passiveButton.OnClick = new();
+
+            if (buttonIsExit)
+                passiveButton.OnClick.AddListener((Action)SplashLogoAnimatorPatch.SceneChanger.ExitGame);
+            else
+                passiveButton.OnClick.AddListener((Action)(() => InfoPopup.Close()));
+        }
+    }
+
+    public static void ShowPopupWithTwoButtons(string message, string firstButtonText, string secondButtonText = "", Action onClickOnFirstButton = null, Action onClickOnSecondButton = null)
+    {
+        if (InfoPopupV2)
+        {
+            var templateExitGame = InfoPopupV2.transform.FindChild("ExitGame");
+            if (!templateExitGame) return;
+
+            var background = InfoPopupV2.transform.FindChild("Background");
+            if (!background) return;
+            background.localScale *= 2f;
+
+            InfoPopupV2.Show(message);
+            templateExitGame.gameObject.SetActive(false);
+            var firstButton = Object.Instantiate(templateExitGame, InfoPopupV2.transform);
+            var secondButton = Object.Instantiate(templateExitGame, InfoPopupV2.transform);
+            
+            if (firstButton)
+            {
+                firstButton.gameObject.SetActive(true);
+                firstButton.name = "FirstButton";
+                var firstButtonTransform = firstButton.transform;
+                firstButton.transform.localPosition = new Vector3(firstButtonTransform.localPosition.x - 1f, firstButtonTransform.localPosition.y - 0.7f, firstButtonTransform.localPosition.z);
+                firstButton.transform.localScale *= 1.2f;
+                var firstButtonGetChild = firstButton.GetChild(0);
+                var textTranslatorTMP = firstButtonGetChild.GetComponent<TextTranslatorTMP>();
+                textTranslatorTMP.TargetText = StringNames.Cancel;
+                textTranslatorTMP.ResetText();
+                textTranslatorTMP.DestroyTranslator();
+                firstButtonGetChild.GetComponent<TextMeshPro>().text = firstButtonText;
+                firstButtonGetChild.GetComponent<TMP_Text>().text = firstButtonText;
+                var passiveButton = firstButton.GetComponent<PassiveButton>();
+                passiveButton.OnClick = new();
+                if (onClickOnFirstButton != null)
+                    passiveButton.OnClick.AddListener((UnityAction)(() => { onClickOnFirstButton(); InfoPopupV2.Close();}));
+                else passiveButton.OnClick.AddListener((UnityAction)(() => InfoPopupV2.Close()));
+            }
+            
+            if (secondButton)
+            {
+                secondButton.gameObject.SetActive(true);
+                secondButton.name = "SecondButton";
+                var secondButtonTransform = secondButton.transform;
+                secondButton.transform.localPosition = new Vector3(secondButtonTransform.localPosition.x + 1f, secondButtonTransform.localPosition.y - 0.7f, secondButtonTransform.localPosition.z);
+                secondButton.transform.localScale *= 1.2f;
+                var secondButtonGetChild = secondButton.GetChild(0);
+                var textTranslatorTMP = secondButtonGetChild.GetComponent<TextTranslatorTMP>();
+                textTranslatorTMP.TargetText = StringNames.Cancel;
+                textTranslatorTMP.ResetText();
+                if (!string.IsNullOrWhiteSpace(secondButtonText))
+                {
+                    textTranslatorTMP.DestroyTranslator();
+                    secondButtonGetChild.GetComponent<TextMeshPro>().text = secondButtonText;
+                    secondButtonGetChild.GetComponent<TMP_Text>().text = secondButtonText;
+                }
+
+                var passiveButton = secondButton.GetComponent<PassiveButton>();
+                passiveButton.OnClick = new();
+                if (onClickOnSecondButton != null)
+                    passiveButton.OnClick.AddListener((UnityAction)(() => { onClickOnSecondButton(); InfoPopupV2.Close(); }));
+                else passiveButton.OnClick.AddListener((UnityAction)(() => InfoPopupV2.Close()));
+            }
         }
     }
 }
